requirements/prod.txt
tensorflow~=2.3.1
torch~=1.7.0
torchvision~=0.8.1
transformers~=3.4.0
datasets~=1.1.2
apache-beam~=2.25.0
Click~=7.1.2
typing_extensions~=3.7.4
nltk~=3.5
sklearn~=0.0
beautifulsoup4~=4.9.3
lxml~=4.6.2
tf-models-official~=2.4.0
unidecode~=1.1.2